Downgrade urllib3 to 1.24.3
Due to urllib3 v1.25.x’s inclusion of `rfc3986` it’s compiling a bunch
of regexes every time `urllib3` is imported… which it is every time
`requests` is imported, which is imported by a ton of Kodi add‐ons.

The best solution here would be either for `urllib3` (or `rfc3986`) to
be smarter about the regex compiling, or for Kodi to not need to
reimport Python libraries on every view change… However, neither of
those are here‐and‐now fixes and upstream `urllib3` have released a
version of 1.24.x which includes the security fix that was previously
only available in the 1.25.x releases, so downgrading to 1.24.3 until
this is handled better either by urllib3, its consumers, or Kodi itself.

Fixes https://github.com/Freso/script.module.urllib3/issues/3
and https://github.com/Freso/script.module.requests/issues/25

See also issue reported to upstream:
https://github.com/urllib3/urllib3/issues/1590

And issue reported to upstream `requests`:
https://github.com/kennethreitz/requests/issues/5080

Based on urllib3-1.24.3.tar.gz from PyPI:
https://pypi.org/project/urllib3/1.24.3/#files